requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==3.2.11
django-filter==21.1
django-taggit==2.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
